requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
twilio>=8.10.0
google-generativeai>=0.3.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
jinja2>=3.1.2
httpx>=0.25.2
pydantic>=2.5.0
greenlet>=3.0.0


